# Refs #28459 -- Improved performance of sql.compiler.cursor_iter().